requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
openai>=1.3.0
python-docx>=0.8.11
PyPDF2>=3.0.0
gunicorn>=20.1.0